add translations and putin components, remove deliveryMode
